# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Charon",
    "description": "3D dependency visualizer for Python projects",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/fitness/validate": {
      "post": {
        "tags": [
          "fitness"
        ],
        "summary": "Validate Fitness Functions",
        "description": "Validate architectural fitness functions against a dependency graph.\n\nThis endpoint evaluates a set of architectural rules against the provided\ndependency graph and returns a detailed validation report.\n\nThe validation can be configured to fail on errors, warnings, or both.\nThis is useful for CI/CD integration where you want to fail builds based\non architectural rule violations.\n\nExample rules:\n- No modules in /api may import from /database directly\n- Maximum efferent coupling: 10\n- No circular dependencies allowed\n- Third-party dependencies must be < 20% of total imports",
        "operationId": "validate_fitness_functions_api_fitness_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FitnessValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitnessValidationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fitness/validate-from-config": {
      "post": {
        "tags": [
          "fitness"
        ],
        "summary": "Validate From Config",
        "description": "Validate using a fitness rule configuration file.\n\nThis endpoint accepts a FitnessRuleConfig object instead of individual rules,\nmaking it easier to maintain rules in a version-controlled configuration file.",
        "operationId": "validate_from_config_api_fitness_validate_from_config_post",
        "parameters": [
          {
            "name": "fail_on_error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Fail On Error"
            }
          },
          {
            "name": "fail_on_warning",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Fail On Warning"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_from_config_api_fitness_validate_from_config_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitnessValidationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fitness/save-result": {
      "post": {
        "tags": [
          "fitness"
        ],
        "summary": "Save Validation Result",
        "description": "Save fitness validation result for historical tracking.\n\nResults are saved to a JSON file in the format:\n{FITNESS_STORAGE_PATH}/{project_name}/fitness_history.jsonl\n\nEach validation result is appended as a new line to enable trend analysis.\nStorage path is configured via FITNESS_STORAGE_PATH environment variable.",
        "operationId": "save_validation_result_api_fitness_save_result_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_-]{0,127}$",
              "description": "Project name (alphanumeric, underscore, hyphen only, max 128 chars)",
              "title": "Project Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FitnessValidationResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fitness/trend/{project_name}": {
      "get": {
        "tags": [
          "fitness"
        ],
        "summary": "Get Fitness Trend",
        "description": "Get historical fitness trend data for a project.\n\nReturns a time series of validation results, optionally filtered by rule_id.\nUseful for tracking architectural health over time and identifying regressions.\nStorage path is configured via FITNESS_STORAGE_PATH environment variable.",
        "operationId": "get_fitness_trend_api_fitness_trend__project_name__get",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_-]{0,127}$",
              "description": "Project name (alphanumeric, underscore, hyphen only, max 128 chars)",
              "title": "Project Name"
            }
          },
          {
            "name": "rule_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitnessTrendResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/fitness/config/example": {
      "get": {
        "tags": [
          "fitness"
        ],
        "summary": "Get Example Config",
        "description": "Get an example fitness rule configuration.\n\nThis returns a sample configuration demonstrating all supported rule types.\nUse this as a starting point for creating your own fitness rules.",
        "operationId": "get_example_config_api_fitness_config_example_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FitnessRuleConfig"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_validate_from_config_api_fitness_validate_from_config_post": {
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/DependencyGraph"
          },
          "global_metrics": {
            "$ref": "#/components/schemas/GlobalMetrics"
          },
          "config": {
            "$ref": "#/components/schemas/FitnessRuleConfig"
          }
        },
        "type": "object",
        "required": [
          "graph",
          "global_metrics",
          "config"
        ],
        "title": "Body_validate_from_config_api_fitness_validate_from_config_post"
      },
...
      "DependencyGraph": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/Node"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/Edge"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "DependencyGraph",
        "description": "Complete dependency graph with metrics."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetails"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail",
          "status_code"
        ],
        "title": "ErrorResponse",
        "description": "Standardized error response payload."
      },
...
      "FitnessRuleConfig": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Config version",
            "default": "1.0"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/FitnessRule"
            },
            "type": "array",
            "title": "Rules",
            "description": "List of fitness rules"
          }
        },
        "type": "object",
        "required": [
          "rules"
        ],
        "title": "FitnessRuleConfig",
        "description": "Configuration file for fitness functions."
      },
...
      "FitnessTrendResponse": {
        "properties": {
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id",
            "description": "Optional filter by rule"
          },
          "data_points": {
            "items": {
              "$ref": "#/components/schemas/FitnessTrendPoint"
            },
            "type": "array",
            "title": "Data Points"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "project_name",
          "data_points",
          "summary"
        ],
        "title": "FitnessTrendResponse",
        "description": "Historical fitness trend data."
      },
      "FitnessValidationRequest": {
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/DependencyGraph",
            "description": "Dependency graph to validate"
          },
          "global_metrics": {
            "$ref": "#/components/schemas/GlobalMetrics",
            "description": "Global metrics"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/FitnessRule"
            },
            "type": "array",
            "title": "Rules",
            "description": "Rules to validate against"
          },
          "fail_on_error": {
            "type": "boolean",
            "title": "Fail On Error",
            "description": "Fail if errors found",
            "default": true
          },
          "fail_on_warning": {
            "type": "boolean",
            "title": "Fail On Warning",
            "description": "Fail if warnings found",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "graph",
          "global_metrics",
          "rules"
        ],
        "title": "FitnessValidationRequest",
        "description": "Request to validate fitness functions."
      },
      "FitnessValidationResult": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed",
            "description": "Whether all rules passed"
          },
          "total_rules": {
            "type": "integer",
            "title": "Total Rules",
            "description": "Total number of rules evaluated"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/FitnessViolation"
            },
            "type": "array",
            "title": "Violations"
          },
          "errors": {
            "type": "integer",
            "title": "Errors",
            "description": "Number of error-level violations",
            "default": 0
          },
          "warnings": {
            "type": "integer",
            "title": "Warnings",
            "description": "Number of warning-level violations",
            "default": 0
          },
          "infos": {
            "type": "integer",
            "title": "Infos",
            "description": "Number of info-level violations",
            "default": 0
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Validation timestamp"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Summary of validation results"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "total_rules",
          "timestamp",
          "summary"
        ],
        "title": "FitnessValidationResult",
        "description": "Result of fitness function validation."
      },
...
      "SaveResultResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the save was successful"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message"
          },
          "file_path": {
            "type": "string",
            "title": "File Path",
            "description": "Path to the saved file"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "file_path"
        ],
        "title": "SaveResultResponse",
        "description": "Response from saving fitness validation result."
      },
...
    }
  }
}